fix ppt unit test
